feature: ordering by Id
When getting objects by Ids and not passing a specific sort: this feature adds (basic) sorting by the given Ids. IMHO it makes sense to return the full objects in order of the given Id's if not specifically passing a sort to the method.

The sorting of the results by the give ids might be refactored for performance later. It feels kind of clunky right now...