only fetch the fields for the getByRef required extra entities...
This adds the fields param to the GetById()-call made in getByRef. This ensures that only the required paths of the entity which are needed for resolution are fetched. As this works top->down, setting fields to the upper-most level, all subdocuments are still fetched. (just that other top level properties aren't sent over the wire)